Fix font familes
Seems that sass is incapable of dealing with unicode (yeah! even in 2020!),
so using '思源黑体' in $family-sans-serif will result in an absolutely incomprehensible
build-time-only error saying "can't resolve ...", which, I presume, is that
processing unicode characters makes webpack to resolve resouces under the
wrong path.
Unfortunately, I was not in the mood of digging deeper into this, the solution
is to remove the unicode characters.
Note that escaping them by \uXXXX will NOT work, since sass does stupid tricks
and removes the preceding backslash.